% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{categories}
\alias{categories}
\title{Table of Codes for Categories}
\format{
A data frame with 37 rows and
2 variables:
\describe{
\item{code}{code of the category consisting of two to four capital letters
or (in one case) digits.}
\item{description}{clear text name of the category}
}
}
\usage{
categories
}
\description{
This dataset contains the codes for many of the race categories for which
data can be obtained from the FIS webpage. Note that due to the structure of
the site, it is not possible to guarantee that the table of code is complete.
There may be other codes that are as of yet missing.
These codes can be used to filter
for a specific category in \code{\link[=query_results]{query_results()}}.
}
\keyword{datasets}
