% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disciplines}
\alias{disciplines}
\title{Table of Codes for Disciplines}
\format{
A data frame with 14 rows and
2 variables:
\describe{
\item{code}{code of the discipline consisting of two or three
(for parasports disciplines) capital letters}
\item{description}{clear text name of the discipline}
}
}
\usage{
disciplines
}
\description{
This dataset contains the codes for all the disciplines for which data
can be obtained from the FIS webpage. These codes can be used to filter
for a specific discipline in \code{\link[=query_athletes]{query_athletes()}}.
}
\keyword{datasets}
