% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisdata-package.R
\docType{package}
\name{fisdata-package}
\alias{fisdata}
\alias{fisdata-package}
\title{fisdata: Query Data From the FIS Web Page}
\description{
Query data from the web site of the International Ski and Snowboard Federation FIS (\url{https://www.fis-ski.com}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stibu81.github.io/fisdata/}
}

}
\author{
\strong{Maintainer}: Stefan Lanz \email{slanz1137@gmail.com}

}
\keyword{internal}
