% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_audio.R
\name{get_athlete_image}
\alias{get_athlete_image}
\title{Get an Image of an Athlete}
\usage{
get_athlete_image(athlete, file = NULL)
}
\arguments{
\item{athlete}{a list or data frame with the field/column \code{competitor_id}
that describe a \emph{single} athlete. The easiest way to create
such a data frame is through the functions \code{\link[=query_athletes]{query_athletes()}},
\code{\link[=query_race]{query_race()}}, or \code{\link[=query_standings]{query_standings()}}. These functions
can return multiple athletes, but \code{get_athlete_image()} only downloads the
image for one athlete. If multiple athletes are passed, only the first
one will be used.}

\item{file}{if given, the image file will be stored to the given path.
The image files are png or jpg and the file ending will be added
automatically depending on the file type. Any user provided file ending
will be ignored.}
}
\value{
\code{NULL} invisibly. As a side effect, the image is displayed and, if \code{file}
is given, stored to the requested path.
}
\description{
Download the image of an athlete from the FIS page. Images are not available
for all athletes.
}
\details{
The image resolution is 100 x 100.
}
