% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_debuts}
\alias{get_debuts}
\title{Get the debut races, podiums or victories for an athlete}
\usage{
get_debuts(
  athlete,
  category = fd_def("category"),
  discipline = fd_def("discipline"),
  by = c("category", "discipline"),
  type = c("race", "podium", "victory")
)
}
\arguments{
\item{athlete}{a list or data frame with fields/columns \code{competitor_id} and
\code{sector} that describe a \emph{single} athlete. The easiest way to create
such a data frame is through the functions \code{\link[=query_athletes]{query_athletes()}},
\code{\link[=query_race]{query_race()}}, or \code{\link[=query_standings]{query_standings()}}. These functions
can return multiple athletes, but \code{query_results()} only returns the
results for one athlete. If multiple athletes are passed, only the first
one will be used.}

\item{category}{abbreviation of the category of the race, e.g., "WC" for
"World Cup". See the dataset \link{categories} for possible values.}

\item{discipline}{abbreviation for the discipline, e.g., "DH" for
"Downhill". See the dataset \link{disciplines} for possible values.}

\item{by}{variables to groups the results by. Possible values are "category",
and "discipline". Values are partially matched. Set this value
to an empty vector (\code{c()}) or \code{NA} to summarise without grouping.}

\item{type}{which type of debut to get: the first race ("race"), the first
podium ("podium"), or the first victory ("victory").}
}
\value{
A tibble with the following columns: \code{athlete}, \code{date}, \code{place}, \code{nation},
\code{sector}, \code{category}, \code{discipline}, \code{rank}, \code{fis_points}, \code{cup_points},
and \code{race_id}.
}
\description{
Get the debut races, podiums or victories for an athlete
}
\examples{
\dontrun{
# get Marco Odermatt's World Cup debuts by discipline
odermatt <- query_athletes("odermatt", "marco")
get_debuts(odermatt, category = "WC")

# get first victory by category
get_debuts(odermatt, by = "category", type = "victory")
}

}
