% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nations}
\alias{nations}
\title{Table of IOC Country Codes}
\format{
A data frame with 223 rows and
3 variables:
\describe{
\item{code}{IOC country code consisting of three capital letters}
\item{country}{name of the country}
\item{current}{is this a code that is currently in use?}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_IOC_country_codes}

Section "Current NOCs" for the countries with \code{current = TRUE} and section
"Historic NOCs and teams, Codes still in use" for those with
\code{current = FALSE}.
}
\usage{
nations
}
\description{
This dataset contains all current and some historic IOC country codes.
These codes can be used to filter for a specific nation in
\code{\link[=query_athletes]{query_athletes()}}.
}
\keyword{datasets}
