% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_audio.R
\name{play_athlete_name}
\alias{play_athlete_name}
\title{Play Athlete's Name in Browser}
\usage{
play_athlete_name(athlete, file = NULL)
}
\arguments{
\item{athlete}{a list or data frame with the field/column \code{competitor_id}
that describe a \emph{single} athlete. The easiest way to create
such a data frame is through the functions \code{\link[=query_athletes]{query_athletes()}},
\code{\link[=query_race]{query_race()}}, or \code{\link[=query_standings]{query_standings()}}. These functions
can return multiple athletes, but \code{play_athlete_name()} only plays the
audio file for one athlete. If multiple athletes are passed, only the first
one will be used.}

\item{file}{if given, the audio file will be stored to the given path.
The file is in the mp3 format and the file ending will be added
automatically. Any user provided file ending will be ignored.}
}
\value{
\code{NULL} invisibly. As a side effect, the mp3 file is played in the browser
(if it exists) and, if \code{file} is given, stored to the requested path.
}
\description{
For some athlete's, an audio file where they speak their name is available
on the FIS page. This functions opens that page in the browser.
}
