% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ranks_over_time}
\alias{plot_ranks_over_time}
\title{Plot athlete's results over time}
\usage{
plot_ranks_over_time(
  results,
  by = c("discipline", "category", "athlete"),
  pos = 1:3,
  interactive = TRUE,
  width = NULL,
  height = NULL
)
}
\description{
Plot athlete's results over time
}
