% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athletes.R
\name{query_athletes}
\alias{query_athletes}
\title{Query Athletes}
\usage{
query_athletes(
  last_name = "",
  first_name = "",
  discipline = "",
  nation = "",
  gender = "",
  birth_year = "",
  brand = "",
  active_only = FALSE
)
}
\arguments{
\item{last_name, first_name}{last and first name. These fields do not support
special characters. For example, use "e" instead of "é" or "è",
"ae" instead of "ä", and "ss" instead of "ß".}

\item{discipline}{abbreviation of the discipline, e.g., AL for alpine skiing}

\item{nation}{abbreviation of the nation, e.g., SUI for Switzerland. The
value is matched exactly.}

\item{gender}{abbreviation of the gender: "M" for male or "F" for female}

\item{birth_year}{birth year. This also supports multiple years separated
by commas (e.g, "1995,1998,2000") or year ranges (e.g., "1990-1995").}

\item{brand}{ski or snowboard brand used by the athlete. This field does not
support special characters (see argument \code{last_name}).}

\item{active_only}{should the query be restricted to active athletes.}
}
\description{
Query athletes using various filters. Omitting a filter means that athletes
with any value in that field will be returned. Filtering is case-insensitive
and string matches are partial, except for \code{nation}.
}
