% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions.R
\name{query_competitions}
\alias{query_competitions}
\title{Query Competitions}
\usage{
query_competitions(event)
}
\arguments{
\item{event}{a list or data frame with fields/columns \code{event_id},
\code{sector} and \code{place} that describe a \emph{single} event. The easiest way to
create such a data frame is through the function \code{\link[=query_events]{query_events()}}.
This function can return multiple events, but \code{query_events()} only
returns the results for one event If multiple events are passed, only
the first one will be used.}
}
\value{
A tibble with the following columns: \code{place}, \code{date}, \code{time}, \code{competition},
\code{sector}, \code{category}, \code{gender}, and \code{race_id}.
}
\description{
Query all the competitions, i.e., individual races, for a given event.
}
\examples{
\dontrun{
# find the Wengen alpine skiing races from the season 2024/2025
wengen_2025 <- query_events(sector = "AL", place = "wengen",
                            category = "WC", season = 2025)

# get all the races that took place during the event
query_competitions(wengen_2025)
}

}
