% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{query_events}
\alias{query_events}
\title{Query Events}
\usage{
query_events(
  selection = c("all", "results", "upcoming"),
  sector = "",
  category = "",
  discipline = "",
  gender = "",
  place = "",
  season = "",
  date = ""
)
}
\arguments{
\item{selection}{which events should be returned: past events, where results
are available ("results") or upcoming events ("upcoming") or both ("all")?}

\item{sector}{abbreviation of the sector, e.g., "AL" for
alpine skiing. See the dataset \link{sectors} for possible values.}

\item{category}{abbreviation of the category of the race, e.g., "WC" for
"World Cup". See the dataset \link{categories} for possible values.}

\item{discipline}{abbreviation for the discipline, e.g., "DH" for
"Downhill". See the dataset \link{disciplines} for possible values.}

\item{gender}{abbreviation of the gender: "M" for male or "F" for female}

\item{place}{location of the race. The API does not
support special characters, but many are handled automatically
(see 'Details').}

\item{season}{year when the season ended, i.e., 2020 stands for the season
2019/2020. It is not possible to filter for multiple seasons at once.}

\item{date}{date at which the event takes place.}
}
\value{
A tibble.
}
\description{
Query Events
}
\details{
The API does not support special character in the field \code{place}.
The following special characters are handled
automatically: à, á, å, ä, æ, ç, ć, č, ð, é, è, ê, ë, ï, ñ, ø, ó, ő, ö,
œ, š, ß, ú, ü, and ž.
Other special characters must be replaced by the suitable
substitute by the user.
}
