% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/races.R
\name{query_race}
\alias{query_race}
\title{Query Race}
\usage{
query_race(result)
}
\arguments{
\item{result}{a list or data frame with fields/columns \code{race_id} and
\code{discipline} that describe a \emph{single} race. The easiest way to create
such a data frame is through the function \code{\link[=query_results]{query_results()}}. This function
can return multiple results, but \code{query_rac()} only returns the
results for one race. If multiple results are passed, only the first
one will be used.}
}
\value{
A tibble with the following columns: \code{rank}, \code{bib}, \code{fis_code}, \code{name},
\code{brand}, \code{birth_year}, \code{nation}, \code{time}, \code{diff_time}, \code{fis_points},
and \code{cup_points}.
}
\description{
Query full results for a race.
}
