% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{query_results}
\alias{query_results}
\title{Query Results for an Athlete}
\usage{
query_results(
  athlete,
  season = fd_def("season"),
  category = fd_def("category"),
  place = "",
  discipline = fd_def("discipline"),
  add_age = TRUE
)
}
\arguments{
\item{athlete}{a list or data frame with fields/columns \code{competitor_id} and
\code{sector} that describe a \emph{single} athlete. The easiest way to create
such a data frame is through the functions \code{\link[=query_athletes]{query_athletes()}},
\code{\link[=query_race]{query_race()}}, or \code{\link[=query_standings]{query_standings()}}. These functions
can return multiple athletes, but \code{query_results()} only returns the
results for one athlete. If multiple athletes are passed, only the first
one will be used.}

\item{season}{year when the season ended, i.e., 2020 stands for the season
2019/2020. It is not possible to filter for multiple seasons at once.}

\item{category}{abbreviation of the category of the race, e.g., "WC" for
"World Cup". Not case-sensitive.
See the dataset \link{categories} for possible values.
If a string not matching a category code is used, a similar string
is searched for in the description column of \link{categories}.}

\item{place}{location of the race. The API does not
support special characters, but many are handled automatically
(see 'Details').}

\item{discipline}{abbreviation for the discipline, e.g., "DH" for
"Downhill". Not case sensitive.
See the dataset \link{disciplines} for possible values.
If a string not matching a discipline code is used, a similar string
is searched for in the description column of \link{disciplines}.}

\item{add_age}{should a column with the athletes age be added
for each race? The age is given in decimal years. Note that for some
athletes, only the birth year is known, in which case the age will be
estimated.}
}
\value{
A tibble with the following columns: \code{athlete}, \code{date}, \code{place}, \code{nation},
\code{sector}, \code{category}, \code{discipline}, \code{rank}, \code{fis_points}, \code{cup_points},
and \code{race_id}. If \code{add_age} is \code{TRUE}, it also contains the column \code{age}
after \code{date}.
}
\description{
Query results for an athlete using various filters. Omitting a filter means
that results with any value in that field will be returned.
Filtering is case-insensitive and for \code{place} string matches are partial.
}
\details{
All filter arguments are set to \code{""} by default. Setting an argument to
\code{""} means that no filtering takes place for this parameter. For those
arguments that have a call to \code{\link[=fd_def]{fd_def()}} as their default value, the default
value can be globally set using \code{\link[=set_fisdata_defaults]{set_fisdata_defaults()}}.

The API does not support special character in the field \code{place}.
The following special characters are handled
automatically: à, á, å, ä, æ, ç, ć, č, ð, é, è, ê, ë, ï, ñ, ø, ó, ő, ö,
œ, š, ß, ú, ü, and ž.
Other special characters must be replaced by the suitable
substitute by the user.

The results are cached such that the same data are only downloaded once
per sessions.
}
\examples{
\dontrun{
# in order to query an athletes results, we first
# have to obtain the competitor id, which is
# required for the query. This can be conveniently
# done with query_athletes().
odermatt <- query_athletes("odermatt", "marco")

# get all of his results
query_results(odermatt)

# get only World Cup Downhill results from the
# season 2023/2024
query_results(
  odermatt,
  category = "WC",
  season = 2024,
  discipline = "DH"
)

# get all results from Kitzbühel. Note that the
# umlaut is removed in the output.
query_results(odermatt, place = "Kitzbühl")
}

}
