% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings.R
\name{query_standings}
\alias{query_standings}
\title{Query Cup Standings for an Athlete}
\usage{
query_standings(sector = "", season = "", category = "", gender = "")
}
\arguments{
\item{sector}{abbreviation of the sector, e.g., "AL" for
alpine skiing. See the dataset \link{sectors} for possible values.}

\item{season}{year when the season ended, i.e., 2020 stands for the season
2019/2020. It is not possible to filter for multiple seasons at once. If
omitted, results are returned for the current season.}

\item{category}{abbreviation of the category for the cup, e.g., "WC" for
"World Cup". See the dataset \link{categories} for possible values; note that
the standing is only available for some of the categories. If an unsupported
category is used, the FIS page unfortunately returns the standings for a
default category.}

\item{gender}{abbreviation of the gender: "M" for male or "F" for female}
}
\description{
Query results for an athlete using various filters. Omitting a filter means
that results with any value in that field will be returned.
Filtering is case-insensitive and for \code{place} string matches are partial.
}
