% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sectors}
\alias{sectors}
\title{Table of Codes for Sectors}
\format{
A data frame with 14 rows and
2 variables:
\describe{
\item{code}{code of the sector consisting of two or three
(for parasports sectors) capital letters}
\item{description}{clear text name of the sector}
}
}
\usage{
sectors
}
\description{
This dataset contains the codes for all the sectors for which data
can be obtained from the FIS webpage. These codes can be used to filter
for a specific sector in \code{\link[=query_athletes]{query_athletes()}}.
}
\keyword{datasets}
