% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{set_fisdata_defaults}
\alias{set_fisdata_defaults}
\alias{get_fisdata_defaults}
\alias{reset_fisdata_defaults}
\alias{fd_def}
\title{Get And Set Default Values for Parameters in Querying Functions}
\usage{
set_fisdata_defaults(
  sector = NULL,
  season = NULL,
  gender = NULL,
  category = NULL,
  discipline = NULL,
  active_only = NULL,
  reset = FALSE
)

get_fisdata_defaults()

reset_fisdata_defaults()

fd_def(
  name = c("sector", "season", "gender", "category", "discipline", "active_only")
)
}
\arguments{
\item{sector}{abbreviation of the sector, e.g., "AL" for
alpine skiing. See the dataset \link{sectors} for possible values.}

\item{season}{year when the season ended, i.e., 2020 stands for the season
2019/2020. It is not possible to filter for multiple seasons at once.}

\item{gender}{abbreviation of the gender: "M" for male/men,
"F" or "W" for female/women.}

\item{category}{abbreviation of the category of the race, e.g., "WC" for
"World Cup". See the dataset \link{categories} for possible values.}

\item{discipline}{abbreviation for the discipline, e.g., "DH" for
"Downhill". See the dataset \link{disciplines} for possible values.}

\item{active_only}{should the query be restricted to active athletes.}

\item{reset}{if \code{TRUE}, the defaults are set to those that are passed to the
function, i.e., all arguments that are omitted are set to \code{""}. If \code{FALSE},
only those defaults that are passed to the function are modified, while
all the others are left unchanged.}

\item{name}{name of the default argument}
}
\description{
\code{set_fisdata_defaults()} sets default values for common arguments that will

\code{reset_fisdata_defaults()} resets all the default values to \code{""}, which is
the same value they get when the package is loaded.

\code{get_fisdata_defaults()} shows the currently selected default values.

\code{fd_def()} returns the current default value for a single parameter; this
function's main use is as default argument in the querying functions.
}
