% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_url}
\alias{show_url}
\title{Show URL associated With a Query}
\usage{
show_url(fisdata_df)
}
\arguments{
\item{fisdata_df}{a table of fisdata-results produced by one of the
\verb{query_*()}-functions.}
}
\value{
a character vector of length one with the URL that was used to create the
table. In an interactive session it also copies the URL to the clipboard
as a side effect.
}
\description{
All the functions \verb{query_*()} call a URL on the FIS web page to collect
their data. \code{show_url()} returns the URL that was used to produce a table
of fisdata-results.
}
\details{
If run from an interactive session, the URL is also copied into the
clipboard such that it can be pasted into a browser.
}
