% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{write_current_defaults}
\alias{write_current_defaults}
\title{Read and Write Defaults from a JSON File}
\usage{
write_current_defaults(file = "~/.fisdata.json", overwrite = FALSE)
}
\arguments{
\item{file}{name of the JSON file to read or write}

\item{overwrite}{should an existing file be overwritten?}
}
\description{
Default settings can be written to a JSON file and read again from
this file. If the file \code{.fisdata.json} exists in the user's home
it is read automatically.
}
