% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{write_defaults}
\alias{write_defaults}
\alias{write_current_defaults}
\title{Read and Write Defaults from a JSON File}
\usage{
write_defaults(
  file = "~/.fisdata.json",
  overwrite = FALSE,
  sector = "",
  season = "",
  gender = "",
  category = "",
  discipline = "",
  active_only = FALSE
)

write_current_defaults(file = "~/.fisdata.json", overwrite = FALSE)
}
\arguments{
\item{file}{name of the JSON file to read or write}

\item{overwrite}{should an existing file be overwritten?}

\item{sector}{abbreviation of the sector, e.g., "AL" for
alpine skiing. Not case-sensitive.
See the dataset \link{sectors} for possible values.
If a string not matching a sector code is used, a similar string
is searched for in the description column of \link{sectors}.}

\item{season}{year when the season ended, i.e., 2020 stands for the season
2019/2020. It is not possible to filter for multiple seasons at once.}

\item{gender}{abbreviation of the gender: "M" for male/men,
"F" or "W" for female/women.}

\item{category}{abbreviation of the category of the race, e.g., "WC" for
"World Cup". Not case-sensitive.
See the dataset \link{categories} for possible values.
If a string not matching a category code is used, a similar string
is searched for in the description column of \link{categories}.}

\item{discipline}{abbreviation for the discipline, e.g., "DH" for
"Downhill". Not case sensitive.
See the dataset \link{disciplines} for possible values.
If a string not matching a discipline code is used, a similar string
is searched for in the description column of \link{disciplines}.}

\item{active_only}{should the query be restricted to active athletes.}
}
\description{
Default settings can be written to a JSON file and read again from
this file. If the file \code{.fisdata.json} exists in the user's home
it is read automatically.
}
