% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{write_fisdata_defaults}
\alias{write_fisdata_defaults}
\alias{write_current_fisdata_defaults}
\alias{read_fisdata_defaults}
\title{Read and Write Defaults from a JSON File}
\usage{
write_fisdata_defaults(
  file = "~/.fisdata.json",
  overwrite = FALSE,
  sector = "",
  season = "",
  gender = "",
  category = "",
  discipline = "",
  active_only = FALSE
)

write_current_fisdata_defaults(file = "~/.fisdata.json", overwrite = FALSE)

read_fisdata_defaults(
  file = "~/.fisdata.json",
  apply = TRUE,
  verbose = !apply || interactive()
)
}
\arguments{
\item{file}{name of the JSON file to read or write}

\item{overwrite}{should an existing file be overwritten?}

\item{sector}{abbreviation of the sector, e.g., "AL" for
alpine skiing. Not case-sensitive.
See the dataset \link{sectors} for possible values.
If a string not matching a sector code is used, a similar string
is searched for in the description column of \link{sectors}.}

\item{season}{year when the season ended, i.e., 2020 stands for the season
2019/2020. It is not possible to filter for multiple seasons at once.}

\item{gender}{abbreviation of the gender: "M" for male/men,
"F" or "W" for female/women.}

\item{category}{abbreviation of the category of the race, e.g., "WC" for
"World Cup". Not case-sensitive.
See the dataset \link{categories} for possible values.
If a string not matching a category code is used, a similar string
is searched for in the description column of \link{categories}.}

\item{discipline}{abbreviation for the discipline, e.g., "DH" for
"Downhill". Not case sensitive.
See the dataset \link{disciplines} for possible values.
If a string not matching a discipline code is used, a similar string
is searched for in the description column of \link{disciplines}.}

\item{active_only}{should the query be restricted to active athletes.}

\item{apply}{should the defaults be applied?}

\item{verbose}{should the function create output. This defaults
to \code{TRUE} in interactive sessions or when \code{apply} is \code{FALSE}.}
}
\value{
\code{write_fisdata_defaults()} and \code{write_current_fisdata_defaults()} return the json-string that
was written to the file (invisibly). \code{read_fisdata_defaults()} returns the default
values that were read as a tibble (invisibly).
}
\description{
Default settings can be written to a JSON file and read again from
this file. If the file \code{.fisdata.json} exists in the user's home
it is read automatically when fisdata is attached in an interactive
session (see 'Details' for how to configure this behaviour).
}
\details{
When fisdata is attached in an interactive session, it tries to load defaults
from a file \code{fisdata.json}. You can use another file by setting the environment
variable \code{FISDATA_DEFAULTS_FILE} to the path to this file before attaching
fisdata. To do this once, you can use \code{\link[=Sys.setenv]{Sys.setenv()}}, to configure R to always
load a different file, you can set \code{FISDATA_DEFAULTS_FILE} in your \code{.Renviron}
file.
}
