% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rank_summary}
\alias{plot_rank_summary}
\title{Plot summary of one or more athlete's results}
\usage{
plot_rank_summary(
  results,
  by = c("category", "discipline"),
  pos = 1:3,
  interactive = TRUE,
  width = NULL,
  height = NULL
)
}
\description{
Plot summary of one or more athlete's results
}
