% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rank_summary}
\alias{plot_rank_summary}
\alias{plot_results_summary}
\title{Plot summary of one or more athlete's results}
\usage{
plot_rank_summary(
  results,
  by = c("category", "discipline"),
  pos = 1:3,
  interactive = TRUE,
  width = NULL,
  height = NULL
)

plot_results_summary(
  results,
  by = c("category", "discipline"),
  variable = c("points", "races", "victories", "podiums", "dnf"),
  relative = FALSE,
  interactive = TRUE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{results}{a tibble with the results of one or more athletes. The tibble
for one athlete can be created with \code{\link[=query_results]{query_results()}} and the tibble for
multiple athletes can be combined with \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}}.}

\item{by}{variables to groups the plots by. Possible values are "category"
(on rows) and "discipline" (on columns). Values are partially matched.
Set this value to an empty vector (\code{c()}) or \code{NA} to create a plot with
a single facet.}

\item{pos}{numeric that controls the summary of ranks. Indicate the
break points for the ranks to summarise. The plot will then show the
counts for the number of ranks that are at least as good as each break
point and worse then the next better break point.}

\item{interactive}{logical indicating whether an interactive plot (with
ggiraph) should be returned. If \code{FALSE}, a static ggplot is returned that
can be turned into an interactive plot with \code{\link[ggiraph:girafe]{ggiraph::girafe()}}.}

\item{width, height}{dimensions for the interactive plot in inches, passed on
to \code{\link[ggiraph:girafe]{ggiraph::girafe()}}. The default values are 6 and 5 inches, respectively.}

\item{variable}{character giving the variable to plot on the y-axis.}

\item{relative}{use relative scale for points, victories, podiums or dnf?}
}
\description{
Plot summary of one or more athlete's results
}
