% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ranks_over_time}
\alias{plot_ranks_over_time}
\title{Plot athlete's results over time}
\usage{
plot_ranks_over_time(
  results,
  by = c("discipline", "category", "athlete"),
  time = c("season", "age"),
  pos = 1:3,
  cumulative = FALSE,
  interactive = TRUE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{results}{a tibble with the results of one or more athletes. The tibble
for one athlete can be created with \code{\link[=query_results]{query_results()}} and the tibble for
multiple athletes can be combined with \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}}.}

\item{by}{variables to groups the plots by. Possible values are "category"
(on rows) and "discipline" (on columns). Values are partially matched.
Set this value to an empty vector (\code{c()}) or \code{NA} to create a plot with
a single facet.}

\item{time}{the variable to use for the x-axis. Possible values are
"season" and "age". The function uses the age at the beginning of the
season for the entire season.}

\item{pos}{numeric that controls the summary of ranks. Indicate one or
multiple break points for the ranks to summarise. The plot will then show
the counts for the number of ranks that are at least as good as each break
point and worse than the next better break point.}

\item{cumulative}{should cumulative counts be plotted?}

\item{interactive}{logical indicating whether an interactive plot (with
ggiraph) should be returned. If \code{FALSE}, a static ggplot is returned that
can be turned into an interactive plot with \code{\link[ggiraph:girafe]{ggiraph::girafe()}}.}

\item{width, height}{dimensions for the interactive plot in inches, passed on
to \code{\link[ggiraph:girafe]{ggiraph::girafe()}}. The default values are 6 and 5 inches, respectively.}
}
\description{
Plot the results of one or several athletes over time.
}
