% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_results_summary}
\alias{plot_results_summary}
\title{Plot summary of one or more athlete's results}
\usage{
plot_results_summary(
  results,
  by = c("category", "discipline"),
  pos = 1:3,
  interactive = TRUE
)
}
\description{
Plot summary of one or more athlete's results
}
