% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athletes.R
\name{query_athletes}
\alias{query_athletes}
\title{Query Athletes}
\usage{
query_athletes(
  last_name = "",
  first_name = "",
  discipline = "",
  nation = "",
  gender = "",
  birth_year = "",
  brand = "",
  active_only = FALSE
)
}
\arguments{
\item{last_name, first_name}{last and first name. The API does not support
special characters, but many are handled automatically (see 'Details').}

\item{discipline}{abbreviation of the discipline, e.g., AL for alpine skiing}

\item{nation}{abbreviation of the nation, e.g., SUI for Switzerland. The
value is matched exactly.}

\item{gender}{abbreviation of the gender: "M" for male or "F" for female}

\item{birth_year}{birth year. This also supports multiple years separated
by commas (e.g, "1995,1998,2000") or year ranges (e.g., "1990-1995").}

\item{brand}{ski or snowboard brand used by the athlete. The API does not
support special characters, but many are handled automatically
(see 'Details').}

\item{active_only}{should the query be restricted to active athletes.}
}
\value{
A tibble with the following columns: \code{active}, \code{fis_code}, \code{name}, \code{nation},
\code{age}, \code{birthdate}, \code{gender}, \code{discipline}, \code{club}, \code{brand}, and
\code{competitor_id}.
}
\description{
Query athletes using various filters. Omitting a filter means that athletes
with any value in that field will be returned. Filtering is case-insensitive
and string matches are partial, except for \code{nation}.
}
\details{
The API does not support special character in the fields \code{last_name},
\code{first_name}, and \code{brand}. The following special characters are handled
automatically: à, á, å, ä, æ, ç, ć, č, ð, é, è, ê, ë, ï, ñ, ø, ó, ő, ö,
œ, š, ß, ú, ü, and ž.
Other special characters must be replaced by the suitable
substitute by the user.
}
