% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athletes.R
\name{query_athletes}
\alias{query_athletes}
\title{Query Athletes}
\usage{
query_athletes(
  last_name = "",
  first_name = "",
  sector = fd_def("sector"),
  nation = "",
  gender = fd_def("gender"),
  birth_year = "",
  brand = "",
  active_only = FALSE
)
}
\arguments{
\item{last_name, first_name}{last and first name. String matching is partial.
The API does not support special characters, but many are handled
automatically (see 'Details').}

\item{sector}{abbreviation of the sector, e.g., "AL" for
alpine skiing. See the dataset \link{sectors} for possible values.}

\item{nation}{abbreviation of the nation, e.g., "SUI" for Switzerland. The
value is matched exactly. See the dataset \link{nations} for possible values.}

\item{gender}{abbreviation of the gender: "M" for male/men,
"F" or "W" for female/women.}

\item{birth_year}{birth year. This also supports multiple years separated
by commas (e.g, "1995,1998,2000") or year ranges (e.g., "1990-1995").}

\item{brand}{ski or snowboard brand used by the athlete. String matching is
partial. The API does not
support special characters, but many are handled automatically
(see 'Details').}

\item{active_only}{should the query be restricted to active athletes.}
}
\value{
A tibble with the following columns: \code{active}, \code{fis_code}, \code{name}, \code{nation},
\code{age}, \code{birthdate}, \code{gender}, \code{sector}, \code{club}, \code{brand}, and
\code{competitor_id}.

\code{active} is a logical indicating whether the athlete is still active. \code{age}
gives the year as an integer, but this value is often missing. \code{birthdate}
is returned as a character.
}
\description{
Query athletes using various filters. Omitting a filter means that athletes
with any value in that field will be returned. Filtering is case-insensitive
and for \code{last_name}, \code{first_name}, and \code{brand}, string matching is partial.
}
\details{
All filter arguments are set to \code{""} by default. Setting an argument to
\code{""} means that no filtering takes place for this parameter. For those
arguments that have a call to \code{\link[=fd_def]{fd_def()}} as their default value, the default
value can be globally set using \code{\link[=set_fisdata_defaults]{set_fisdata_defaults()}}.

The API does not support special character in the fields \code{last_name},
\code{first_name}, and \code{brand}. The following special characters are handled
automatically: à, á, å, ä, æ, ç, ć, č, ð, é, è, ê, ë, ï, ñ, ø, ó, ő, ö,
œ, š, ß, ú, ü, and ž.
Other special characters must be replaced by the suitable
substitute by the user.

One use of this function is to get the competitor id for an athlete, which
is needed in order to query an athletes results with \code{\link[=query_results]{query_results()}}.

The results are cached such that the same data are only downloaded once
per sessions.
}
\examples{
\dontrun{
# find Swiss athletes with last name "Cuche"
query_athletes("cuche", nation = "SUI")

# find French alpine skiers using Rossignol skis
query_athletes(
  sector = "AL",
  nation = "FRA",
  brand = "Rossignol",
  active_only = TRUE
)

# find Loïc Maillard. Note that even if the "ï" may be used in the query,
# the name the name is returned without the special character.
query_athletes("meillard", "loïc")

# the query works the same without the special character
query_athletes("meillard", "loic")
}

}
