% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitions.R
\name{query_competitions}
\alias{query_competitions}
\title{Query Competitions}
\usage{
query_competitions(event)
}
\arguments{
\item{event}{a list or data frame with fields/columns \code{event_id} and
\code{sector} that describe a \emph{single} event. The easiest way to create
such a data frame is through the function \code{\link[=query_events]{query_events()}}. This function
can return multiple events, but \code{query_events()} only returns the
results for one event If multiple events are passed, only the first
one will be used.}
}
\description{
Query Competitions
}
