% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{query_results}
\alias{query_results}
\title{Query Results for an Athlete}
\usage{
query_results(athlete, season = "", category = "", place = "", discipline = "")
}
\arguments{
\item{athlete}{a list or data frame with fields/columns \code{competitor_id} and
\code{sector} that describe a \emph{single} athlete. The easiest way to create
such a data frame is through the function \code{\link[=query_athletes]{query_athletes()}}. This function
can return multiple athletes, but \code{query_results()} only returns the
results for one athlete. If multiple athletes are passed, only the first
one will be used.}

\item{season}{year when the season ended, i.e., 2020 stands for the season
2019/2020. It is not possible to filter for multiple seasons at once.}

\item{category}{abbreviation of the category of the race, e.g., "WC" for
"World Cup". See the dataset \link{categories} for possible values.}

\item{place}{location of the race. The API does not
support special characters, but many are handled automatically
(see 'Details').}

\item{discipline}{abbreviation for the discipline, e.g., "DH" for
"Downhill".}
}
\value{
A tibble with the following columns: \code{athlete}, \code{date}, \code{place}, \code{nation},
\code{sector}, \code{category}, \code{discipline}, \code{rank}, \code{fis_points}, \code{cup_points},
and \code{race_id}.
}
\description{
Query results for an athlete using various filters. Omitting a filter means
that results with any value in that field will be returned.
Filtering is case-insensitive and for \code{place} string matches are partial.
}
\details{
The API does not support special character in the field \code{place}.
The following special characters are handled
automatically: à, á, å, ä, æ, ç, ć, č, ð, é, è, ê, ë, ï, ñ, ø, ó, ő, ö,
œ, š, ß, ú, ü, and ž.
Other special characters must be replaced by the suitable
substitute by the user.
}
